requirements.txt
paho-mqtt 

